MOD : Remove all pylint alerts on scripts folder
